refactor: Move `NoteroItem` into new `models` directory
